backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.0
pandas==2.1.3
email-validator==2.1.0
flask-jwt-extended==4.6.0
gunicorn==21.2.0
psycopg2-binary==2.9.9



